requirements-minimal.txt
celery[redis]==5.1.2
fastapi==0.75.0
ocflib
paramiko==2.7.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-keycloak==0.22.0
redis==3.5.3
requests==2.24.0
uvicorn==0.12.2
